requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
apscheduler
watchdog
streamlit-autorefresh